pyproject.toml
[project]
name = "confluence-exporter"
version = "0.1.0"
description = "递归导出 Confluence 页面及其子页面 (HTML / Markdown)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "atlassian-python-api>=3.41.0",
    "click>=8.1.0",
    "markdownify>=0.11.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
]

[project.scripts]
confluence-export = "confluence_exporter.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/confluence_exporter"]
